requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.2.2
sumy==0.11.0
PyYAML==6.0.2
numpy>=1.26,<3
nltk==3.8.1
